pyproject.toml
[tool.poetry]
name = "integrations-testing-framework"
version = "0.2.0"
description = ""
authors = ["y42"]

[tool.poetry.dependencies]
python = "^3.9"
requests = ">2.20.0"
responses = "0.17.0"
vcrpy = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
